starting work on update route
This commit introduces the routing package and puts
the creation of the http server and the registering
and definition of handlers into that package.

Started on the update route function, which I will
do a bit differently compared to the online course
from udemy I'm following. Having some trouble
getting cgo to compile on windows, so I'll just move
to my linux machine for the time being